% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_compile_stan_functions.R
\name{csm_compile_stan_functions}
\alias{csm_compile_stan_functions}
\title{Compile functions defined in a Stan source file}
\usage{
csm_compile_stan_functions(stan_file)
}
\arguments{
\item{stan_file}{a length-one character vector
with the path to a Stan source file}
}
\description{
Compile functions defined in a Stan source file
}
