% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_create_stan_dydt.R
\name{csm_create_stan_dydt}
\alias{csm_create_stan_dydt}
\title{Create a Stan function for rate of change of state variables}
\usage{
csm_create_stan_dydt(name, state, parameters = NULL, forcings = NULL)
}
\arguments{
\item{name}{name of the resulting Stan function}

\item{state}{a list vector containing CSM state
variables defined using \code{\link[csmdeveloper:csm_create_state]{csmdeveloper::csm_create_state()}}
in the intended order}

\item{parameters}{an optional list vector containing
CSM parameters defined using \code{\link[csmdeveloper:csm_create_parameter]{csmdeveloper::csm_create_parameter()}}
the names of all parameters}

\item{forcings}{an optional character vector
containing the names of all forcing variables}
}
\description{
Create a Stan function for rate of change of state variables
}
