% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_create_stan_dydt.R
\name{csm_create_stan_dydt}
\alias{csm_create_stan_dydt}
\title{Create a Stan function for rate of change of state variables}
\usage{
csm_create_stan_dydt(name, state, arg_names = c(state = "y"), ...)
}
\arguments{
\item{name}{name of the resulting Stan function}

\item{arg_names}{a named character vector whose elements
correspond to the argument name in the function to be
generated and whose names correspond to either \code{states}
or the name of the argument in \code{...}}

\item{...}{optional arguments of list vectors containing
CSM parameters defined using \code{\link[csmdeveloper:csm_create_parameter]{csmdeveloper::csm_create_parameter()}}
or CSM variables defined using \code{\link[csmdeveloper:csm_create_variable]{csmdeveloper::csm_create_variable()}}}

\item{states}{a list vector containing CSM state
variables defined using \code{\link[csmdeveloper:csm_create_state]{csmdeveloper::csm_create_state()}}
in the intended order}
}
\description{
Create a Stan function for rate of change of state variables
}
